Working Chapter and Manager but still issue on index

+++ Stories/JACoverCollectionViewController.h
@@ -9,7 +9,6 @@
 #import <UIKit/UIKit.h>
 #import "JACoverCollectionViewCell.h"
 #import "JALoaderView.h"
-#import "JAManagerData.h"
 #import "JAStoriesModel.h"
 #import "JAStorieModel.h"
 #import "JACoverSeguePop.h"
@@ -26,5 +25,5 @@ typedef NS_ENUM(NSInteger, JAAnimDirection) {
 @property (nonatomic,strong) JAManagerData *manager;
 @property (strong, nonatomic) JALoaderView *loaderView;
 @property (strong, nonatomic) UILabel *nameViewLBL;
-@property int currentIndex;
+@property NSInteger currentIndex;
 @end
